requirements.txt
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
langchain==0.2.14
langchain-community==0.2.12
python-dotenv==1.0.1
openai==1.47.1
pandas==2.2.2
pyyaml==6.0.2
tqdm==4.66.4
tabulate==0.9.0
groq>=0.11.0